requirements.txt
django
djangorestframework
drf-yasg
django-jazzmin
python-dotenv
djangorestframework-simplejwt
